+++ public/index.html
@@ -11,7 +11,7 @@
 
 <body>
   <div id="application"></div>
-  <script src="./main.js"></script>
+  <script src="./main.js" type="module"></script>
 </body>
 
 </html>
