+++ public/main.js
@@ -1,3 +1,5 @@
+import createLinks from "./modules/create_links.js";
+
 const dataForMainPage = {
   login: {
     name: 'Вход',
@@ -15,24 +17,6 @@ const dataForMainPage = {
     cl: 'main_link',
   },
 };
-
-function createLinks(data = {
-  by_default: {
-    name: 'undefined',
-    link: 'undefined',
-    cl: 'main_link'
-  } }, parent = document.body) {
-  parent.innerHTML = '';
-  Object.values(data).forEach(function (item) {
-    const pageItem = document.createElement('a');
-    pageItem.textContent = item.name;
-    pageItem.href = `/${item.link}`;
-    pageItem.dataset.section = item.link;
-    pageItem.classList.add(item.cl);
-
-    parent.appendChild(pageItem);
-  });
-}
 
 function createMainPage() {
   let app = document.getElementById("application")
